Add files via upload
This project aims to provide a comprehensive overview of the data preprocessing steps, model development process, and evaluation results for a machine learning project. The project Sentiment Analysis is a data science project that involves the use of machine learning techniques to analyze and classify textual data based on the sentiment expressed. The project aims to build a predictive model capable of determining whether a given text conveys positive, negative, or neutral sentiment.